using unsigned int

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -13,7 +13,8 @@
 int _atoi(char *s)
 {
 	int i = 0, stat = 0, first, last, len = strlen(s), j = 1;
-	unsigned int res = 0;
+
+	double int res = 0;
 
 	for (i = 0; i <= len; i++)
 	{
